requirements.txt
asgiref==3.9.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2.4
fonttools==4.58.4
google-genai==1.26.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.0
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
scipy==1.16.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
gunicorn==23.0.0
